split launcher into multiple files
